+++ tools/lib/tests/test_logreader.py
@@ -4,16 +4,15 @@ import numpy as np
 import unittest
 from parameterized import parameterized
 import requests
-
+from openpilot.tools.lib.logreader import LogReader, parse_indirect, parse_slice, ReadMode
 from openpilot.tools.lib.route import SegmentRange
-from openpilot.tools.lib.srreader import ReadMode, SegmentRangeReader, parse_slice, parse_indirect
 
 NUM_SEGS = 17 # number of segments in the test route
 ALL_SEGS = list(np.arange(NUM_SEGS))
 TEST_ROUTE = "344c5c15b34f2d8a/2024-01-03--09-37-12"
 QLOG_FILE = "https://commadataci.blob.core.windows.net/openpilotci/0375fdf7b1ce594d/2019-06-13--08-32-25/3/qlog.bz2"
 
-class TestSegmentRangeReader(unittest.TestCase):
+class TestLogReader(unittest.TestCase):
   @parameterized.expand([
     (f"{TEST_ROUTE}", ALL_SEGS),
     (f"{TEST_ROUTE.replace('/', '|')}", ALL_SEGS),
@@ -38,7 +37,6 @@ class TestSegmentRangeReader(unittest.TestCase):
     (f"https://useradmin.comma.ai/?onebox={TEST_ROUTE.replace('/', '|')}", ALL_SEGS),
     (f"https://useradmin.comma.ai/?onebox={TEST_ROUTE.replace('/', '%7C')}", ALL_SEGS),
     (f"https://cabana.comma.ai/?route={TEST_ROUTE}", ALL_SEGS),
-    (f"cd:/{TEST_ROUTE}", ALL_SEGS),
   ])
   def test_indirect_parsing(self, identifier, expected):
     parsed, _, _ = parse_indirect(identifier)
@@ -54,7 +52,7 @@ class TestSegmentRangeReader(unittest.TestCase):
         shutil.copyfileobj(r.raw, f)
 
     for f in [QLOG_FILE, qlog.name]:
-      l = len(list(SegmentRangeReader(f)))
+      l = len(list(LogReader(f)))
       self.assertGreater(l, 100)
 
   @parameterized.expand([
@@ -72,14 +70,14 @@ class TestSegmentRangeReader(unittest.TestCase):
       parse_slice(sr)
 
   def test_modes(self):
-    qlog_len = len(list(SegmentRangeReader(f"{TEST_ROUTE}/0", ReadMode.QLOG)))
-    rlog_len = len(list(SegmentRangeReader(f"{TEST_ROUTE}/0", ReadMode.RLOG)))
+    qlog_len = len(list(LogReader(f"{TEST_ROUTE}/0", ReadMode.QLOG)))
+    rlog_len = len(list(LogReader(f"{TEST_ROUTE}/0", ReadMode.RLOG)))
 
     self.assertLess(qlog_len * 6, rlog_len)
 
   def test_modes_from_name(self):
-    qlog_len = len(list(SegmentRangeReader(f"{TEST_ROUTE}/0/q")))
-    rlog_len = len(list(SegmentRangeReader(f"{TEST_ROUTE}/0/r")))
+    qlog_len = len(list(LogReader(f"{TEST_ROUTE}/0/q")))
+    rlog_len = len(list(LogReader(f"{TEST_ROUTE}/0/r")))
 
     self.assertLess(qlog_len * 6, rlog_len)
